move db routing e2e to a new file, make tests nicer, add test coverage for not showing destination databases outside of database settings
